Create separate component for name input
